+++ encode.go
@@ -2,7 +2,6 @@ package pq
 
 import (
 	"bytes"
-	"database/sql"
 	"encoding/binary"
 	"encoding/hex"
 	"errors"
@@ -603,10 +602,3 @@ func encodeBytea(v []byte) (result []byte) {
 	hex.Encode(result[2:], v)
 	return result
 }
-
-// NullTime represents a [time.Time] that may be null.
-// NullTime implements the [sql.Scanner] interface so
-// it can be used as a scan destination, similar to [sql.NullString].
-//
-// Deprecated: this is an alias for [sql.NullTime].
-type NullTime = sql.NullTime