+++ CHANGELOG.md
@@ -6,12 +6,15 @@ unreleased
 - Support protocol 3.2, and the `min_protocol_version` and
   `max_protocol_version` DSN parameters ([#1258]).
 
+- Support `sslmode=prefer` and `sslmode=allow` ([#1270]).
+
 ### Fixes
 
 - Fix SSL key permission check to allow modes stricter than 0600/0640#1265 ([#1265]).
 
 [#1258]: https://github.com/lib/pq/pull/1258
 [#1265]: https://github.com/lib/pq/pull/1265
+[#1270]: https://github.com/lib/pq/pull/1270
 
 v1.11.2 (2026-02-10)
 --------------------